Reorganizing code and improving logging
