--- utils/log.go
@@ -1,19 +1,0 @@
-package utils
-
-import (
-	"log"
-	"os"
-)
-
-var (
-	Log *log.Logger
-)
-
-func NewLog(logpath string) {
-	println("LogFile: " + logpath)
-	file, err := os.Create(logpath)
-	if err != nil {
-		panic(err)
-	}
-	Log = log.New(file, "", log.LstdFlags|log.Lshortfile)
-}
